refactor: . move field rendering to TaskFieldRenderer.ts
